requirements.txt
asgiref==3.4.1
astroid==2.8.6
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.5
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.0
platformdirs==2.4.1
pylint==2.11.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.6
wheel==0.37.1
wrapt==1.13.3
django-environ==0.8.1
